fix(skills): lint_vault and its test split below the size cap again; log fields are paths only when they read as one
Three findings from the Codex review's nineteenth pass on #182:

- lint_vault.py had grown back to 720 lines. The link graph and Raw Source
  provenance move to vault_links.py (201 lines), is_raw_source joins the
  frontmatter module, and lint_vault.py is under 550 lines.
- test_lint_vault.py had reached 725 lines. The fixture vault, constants, and
  subprocess runner move to tests/skills/lint_vault_support.py with the
  `vault` fixture in conftest.py; the link-graph and provenance cases move to
  test_vault_links.py, the rest stays in test_lint_vault.py.
- A `schema | updated SCHEMA.md` entry was reported as a missing file. A log
  field is a path only when it reads as one — no whitespace before its first
  `/` — and `schema` joins `query` as a free-text operation. A root note with
  spaces is logged as `./My Root Note.md`, which the test covers.
