+++ skills/core/maintaining-obsidian/references/lint_vault.py
@@ -41,8 +41,10 @@ the auditor reads before acting on:
      body, and a typed note hashes the one Raw Source note its body wikilinks.
      A typed note with neither is `unresolved`, whether or not it stores a
      digest: its original is remote, and nothing in the vault stands in for it.
-  5. `log.md` entries naming files that do not exist anywhere in the vault
-     (a `query` entry's fields are a question, not a path, and are not checked).
+  5. `log.md` entries naming files that do not exist anywhere in the vault. A
+     field counts as a path when it ends in an extension and has no whitespace
+     before its first `/` (`Wiki/My Note.md`; a root note with spaces is logged
+     as `./My Note.md`); `query` and `schema` entries are free text, not checked.
   6. Frontmatter tag counts, each marked `declared` when the tag or its top-level
      segment is a backticked list item under SCHEMA.md's `## Tag Taxonomy`;
      `--tag-min` sets `exceeds` for undeclared tags only.
@@ -72,7 +74,6 @@ Output: JSON (--json) or a short text summary.
 from __future__ import annotations
 
 import argparse
-import hashlib
 import json
 import re
 import sys
@@ -82,15 +83,16 @@ from pathlib import Path
 
 from vault_frontmatter import (
     is_empty,
+    is_raw_source,
     note_tags,
     note_type,
     parse_frontmatter,
     read_text,
     split_frontmatter,
-    strip_code,
     text_lines,
     yaml_fences,
 )
+from vault_links import link_facts, raw_source_facts
 
 NON_NOTE_ROOT_FILES = {"AGENTS.md", "SCHEMA.md", "CLAUDE.md", "README.md", "index.md", "log.md"}
 # The standard audit-report folder: every run writes a new report there, so
@@ -100,27 +102,25 @@ LOG_FILE = "log.md"
 SCHEMA_FILE = "SCHEMA.md"
 DUP_CANDIDATE_FLOOR = 0.6
 
-# A link target with an explicit extension other than .md is an attachment
-# embed, not a note relationship — unless a note of that exact name exists
-# (`[[Node.js]]` is a note when Node.js.md is).
-EXTENSION_RE = re.compile(r"\.[A-Za-z0-9]+$")
-# A log token (one pipe-delimited field) names a file when it ends in an
-# extension that starts with a letter (`Raw/recording.mp3`, `Wiki/My Note.md`;
-# not `bump to v1.2`).
+# A log field names a file when it ends in a letter-led extension and reads as
+# a path: no whitespace at all, or none before its first `/` (`Raw/recording.mp3`,
+# `Wiki/My Note.md`; not `bump to v1.2`, not `updated SCHEMA.md`).
 FILE_TOKEN_RE = re.compile(r"^\S.*\.[A-Za-z][A-Za-z0-9]{0,9}$")
-WIKILINK_RE = re.compile(r"\[\[([^\]|#]+)(?:#[^\]|]*)?(?:\|[^\]]*)?\]\]")
 TYPE_NAME_RE = re.compile(r"^[A-Za-z0-9_-]+$")
 TAXONOMY_HEADING_RE = re.compile(r"^##\s+.*\btaxonomy\b", re.IGNORECASE)
 TAXONOMY_ITEM_RE = re.compile(r"^\s*[-*]\s+`#?([^`\s]+)`")
 LOG_ENTRY_RE = re.compile(r"^\s*(?:#+\s*|-\s*)?\[\d{4}-\d{2}-\d{2}\]\s*([A-Za-z-]+)?")
 # Log entries whose fields are free text rather than a path (a question may end
 # in a filename without naming a file to check).
-FREE_TEXT_OPS = {"query"}
+FREE_TEXT_OPS = {"query", "schema"}
 
 
-def is_raw_source(fm: dict | None) -> bool:
-    """A Raw Source note to the script: carries `sha256` and no `type:`."""
-    return fm is not None and "sha256" in fm and note_type(fm) is None
+def is_path_token(token: str) -> bool:
+    """See FILE_TOKEN_RE: a letter-led extension, and no whitespace before the first `/`."""
+    if not FILE_TOKEN_RE.match(token):
+        return False
+    head = token.split("/", 1)[0] if "/" in token else token
+    return re.search(r"\s", head) is None
 
 
 def exceeds(value: float, threshold: float | None) -> bool | None:
@@ -331,177 +331,6 @@ def type_facts(scoped: list[dict], declared, field_empty_pct, undeclared_pct) ->
     return types, untyped
 
 
-def _frontmatter_values(value) -> list[str]:
-    """Every string inside a parsed frontmatter value (scalar, list, or mapping)."""
-    if isinstance(value, str):
-        return [value]
-    if isinstance(value, list):
-        return [s for item in value for s in _frontmatter_values(item)]
-    if isinstance(value, dict):
-        return [s for item in value.values() for s in _frontmatter_values(item)]
-    return []
-
-
-def link_text(note: dict) -> str:
-    """What the graph scans: the body outside code, plus the parsed frontmatter
-    values — a `[[link]]` in a YAML comment is not a link."""
-    values = _frontmatter_values(note["fm"] or {})
-    return strip_code(note["body"]) + "\n" + "\n".join(values)
-
-
-def link_facts(notes: list[dict], scoped: list[dict], raw: list[dict]) -> dict:
-    """Link graph over the wiki-layer `notes`; a link to one of the `raw` captures is
-    provenance, not a relationship, and counts for neither side."""
-    by_rel = {note["rel"] for note in notes}
-    by_stem: dict[str, list[str]] = defaultdict(list)
-    for note in notes:
-        by_stem[Path(note["rel"]).stem.lower()].append(note["rel"])
-    raw_rels = {note["rel"] for note in raw}
-    raw_stems = {Path(note["rel"]).stem.lower() for note in raw}
-
-    def resolve(target: str, source: str = "") -> str | None:
-        # A link with a path resolves by that path only. A bare name resolves by
-        # basename; when several notes share it, the one in the source note's
-        # own folder wins, and otherwise the link stays unresolved rather than
-        # being handed to whichever note sorts first (search-strategies.md:
-        # an ambiguous match is left unresolved).
-        name = target[:-3] if target.endswith(".md") else target
-        if "/" in name:
-            return f"{name}.md" if f"{name}.md" in by_rel else None
-        hits = by_stem.get(name.lower(), [])
-        if len(hits) == 1:
-            return hits[0]
-        same_folder = [rel for rel in hits if Path(rel).parent == Path(source).parent]
-        return same_folder[0] if len(same_folder) == 1 else None
-
-    def is_attachment(target: str, source: str) -> bool:
-        ext = EXTENSION_RE.search(target)
-        return bool(ext) and ext.group(0).lower() != ".md" and resolve(target, source) is None
-
-    def is_raw_capture(target: str, source: str) -> bool:
-        # Provenance is a link that can only mean a capture: by path, or by a
-        # bare name no wiki note has. An ambiguous wiki name stays a wiki link.
-        name = target[:-3] if target.endswith(".md") else target
-        if "/" in name:
-            return f"{name}.md" in raw_rels
-        return name.lower() in raw_stems and name.lower() not in by_stem
-
-    outbound: dict[str, int] = {}
-    inbound: Counter = Counter()
-    for note in notes:
-        targets = set()
-        for match in WIKILINK_RE.finditer(link_text(note)):
-            target = match.group(1).strip()
-            if (
-                target
-                and not is_attachment(target, note["rel"])
-                and not is_raw_capture(target, note["rel"])
-            ):
-                targets.add(target)
-        resolved = {resolve(target, note["rel"]) for target in targets}
-        # A link to the note itself, under any spelling, is not an outbound edge.
-        outbound[note["rel"]] = sum(1 for target in targets if resolve(target, note["rel"]) != note["rel"])
-        for rel in resolved - {None, note["rel"]}:
-            inbound[rel] += 1
-
-    per_note = {
-        note["rel"]: {"inbound": inbound[note["rel"]], "outbound": outbound[note["rel"]]}
-        for note in scoped
-    }
-    orphans = sorted(
-        rel for rel, counts in per_note.items() if not counts["inbound"] and not counts["outbound"]
-    )
-    return {"orphans": orphans, "notes": per_note}
-
-
-def sha256_body(body: str) -> str:
-    return hashlib.sha256(body.encode("utf-8")).hexdigest()
-
-
-def sha256_file(path: Path) -> str:
-    """A note's body digest per raw-sources.md; a non-note file's raw bytes."""
-    if path.suffix == ".md":
-        return sha256_body(split_frontmatter(read_text(path))[1])
-    return hashlib.sha256(path.read_bytes()).hexdigest()
-
-
-def _source_url_file(note: dict, files: dict[str, Path]) -> Path | None:
-    """The vault file `source_url` names, when it names one (not a URL, not the note itself)."""
-    source_url = note["fm"].get("source_url")
-    if not isinstance(source_url, str) or not source_url or "://" in source_url:
-        return None
-    candidate = files.get(source_url.lstrip("/")) or files.get(
-        (Path(note["rel"]).parent / source_url).as_posix()
-    )
-    return candidate if candidate is not None and candidate != note["path"] else None
-
-
-def _linked_raw_source(note: dict, files: dict[str, Path], md_by_stem: dict) -> Path | None:
-    """The Raw Source note (`sha256`, no `type:`) the body wikilinks — when it links exactly one.
-    A link with a path resolves by that path only; a bare name by basename, and only
-    when every note of that name is a capture (a wiki note of the same name makes
-    the link a wiki relationship, not provenance)."""
-
-    def frontmatter_of(rel: str) -> dict | None:
-        fm_text, _ = split_frontmatter(read_text(files[rel]))
-        return parse_frontmatter(fm_text) if fm_text is not None else None
-
-    hits: set[Path] = set()
-    for match in WIKILINK_RE.finditer(strip_code(note["body"])):
-        target = match.group(1).strip()
-        name = target[:-3] if target.endswith(".md") else target
-        if "/" in name:
-            rels = [f"{name}.md"] if f"{name}.md" in files else []
-        else:
-            rels = md_by_stem.get(name.lower(), [])
-            if not all(is_raw_source(frontmatter_of(rel)) for rel in rels):
-                continue
-        for rel in rels:
-            if is_raw_source(frontmatter_of(rel)):
-                hits.add(files[rel])
-    return hits.pop() if len(hits) == 1 else None
-
-
-def raw_source_facts(scoped: list[dict], vault: Path, files: dict[str, Path]) -> list[dict]:
-    md_by_stem: dict[str, list[str]] = defaultdict(list)
-    for rel in files:
-        if rel.endswith(".md"):
-            md_by_stem[Path(rel).stem.lower()].append(rel)
-    out = []
-    for note in scoped:
-        fm = note["fm"]
-        if fm is None or "sha256" not in fm:
-            continue
-        stored = fm["sha256"] if isinstance(fm["sha256"], str) else ""
-        # The provenance pair: a typed note's digest is of the Raw Source it was
-        # ingested from, never of its own synthesized body.
-        target = _source_url_file(note, files)
-        if target is None and note_type(fm) is not None:
-            target = _linked_raw_source(note, files, md_by_stem)
-        if target is not None:
-            hashed_file, recomputed = target.relative_to(vault).as_posix(), sha256_file(target)
-        elif note_type(fm) is None:
-            hashed_file, recomputed = note["rel"], sha256_body(note["body"])
-        else:
-            hashed_file, recomputed = None, None
-        if recomputed is None:
-            status = "unresolved"
-        elif not stored:
-            status = "unhashed"
-        else:
-            status = "fresh" if stored == recomputed else "drift"
-        out.append(
-            {
-                "path": note["rel"],
-                "hashed_file": hashed_file,
-                "status": status,
-                "stored": stored,
-                "recomputed": recomputed,
-            }
-        )
-    return out
-
-
 def log_facts(vault: Path, files: dict[str, Path]) -> dict:
     log = vault / LOG_FILE
     if not log.is_file():
@@ -517,7 +346,7 @@ def log_facts(vault: Path, files: dict[str, Path]) -> dict:
                 continue
         for part in line.split("|"):
             token = part.strip().strip("`")
-            if not FILE_TOKEN_RE.match(token):
+            if not is_path_token(token):
                 continue
             if token in files or token.lstrip("./") in files:
                 continue
