+++ tests/skills/test_lint_vault.py
@@ -1,14 +1,10 @@
-"""Contract test for maintaining-obsidian's LINT script.
+"""Contract test for maintaining-obsidian's LINT script — schema, scope, tags, log, titles, thresholds.
 
-Builds a tiny vault in tmp_path, runs `references/lint_vault.py --json` the way
-the skill runs it (a subprocess), and asserts the facts it reports: block-list
-frontmatter reads as filled, undeclared fields are counted against SCHEMA.md,
-the orphan graph, sha256 drift / unhashed / unresolved provenance, log entries
-naming missing files, tag counts against the taxonomy, duplicate-title
-candidates, and the threshold flags.
+Runs `references/lint_vault.py --json` the way the skill runs it (a subprocess) on
+the fixture vault from lint_vault_support and asserts the facts it reports. The
+link graph and Raw Source provenance are covered by test_vault_links.py.
 """
 
-import hashlib
 import json
 import os
 import subprocess
@@ -17,132 +13,16 @@ from pathlib import Path
 
 import pytest
 
-PROJECT_ROOT = Path(__file__).resolve().parents[2]
-SCRIPT = PROJECT_ROOT / "skills" / "core" / "maintaining-obsidian" / "references" / "lint_vault.py"
-
-SCHEMA = """---
-type: schema
-created: 2026-05-06
----
-
-# Test vault — schema
-
-## Universal Frontmatter
-
-```yaml
----
-type: source | entity
-created: YYYY-MM-DD
-tags: []
----
-```
-
-## Source
-
-```yaml
----
-type: source
-source_url: ""
-source_author: ""          # author or organization
-sha256: ""
----
-```
-
-## Entity
-
-```yaml
----
-type: entity
-aliases: []
----
-```
-
-## Tag Taxonomy
-
-- `arcforge` — project tag
-- `entity` — entity notes (`entity/tool`, ...)
-
-LINT checks:
-- Unknown top-level tags → flag.
-
-## Audit Thresholds
-
-- Field empty in 90%+ of a type → EVOLVE candidate (`--field-empty-pct 90`).
-"""
-
-ALPHA_BODY = "# Alpha\n\nSee [[alpha-note-draft|the draft]] and ![[diagram.png]].\n"
-ALPHA = (
-    "---\n"
-    "type: source\n"
-    "created: 2026-05-06\n"
-    "tags:\n"
-    "  - arcforge\n"
-    "  - tdd\n"
-    'source_url: "https://example.com/alpha"\n'
-    'source_author: ""\n'
-    f'sha256: "{"0" * 64}"\n'
-    "---\n"
-    f"{ALPHA_BODY}"
+from .lint_vault_support import (
+    SCRIPT,
+    SCHEMA,
+    GAMMA,
+    PRESETS,
+    GENERIC_SOURCE,
+    PAPER_SOURCE,
+    _run,
+    _pair,
 )
-
-DRAFT = """---
-type: source
-created: 2026-05-07
-tags: [arcforge]
-source_url: "https://example.com/alpha-draft"
-sha256: ""
----
-# Alpha (draft)
-
-Supersedes [[Wiki/alpha-note]].
-"""
-
-GAMMA = """---
-type: entity
-created: 2026-05-08
-tags: [entity, entity/tool, tdd]
-aliases: []
-extra_field: yes
----
-# Gamma
-
-Nothing links here and this links nowhere.
-"""
-
-LOG = """# Log
-
-## [2026-05-06] create | source | Wiki/alpha-note.md
-## [2026-05-07] create | entity | Wiki/deleted-note.md
-## [2026-05-08] query | how does alpha work?
-"""
-
-
-@pytest.fixture
-def vault(tmp_path: Path) -> Path:
-    (tmp_path / "SCHEMA.md").write_text(SCHEMA, encoding="utf-8")
-    (tmp_path / "AGENTS.md").write_text("# Contract\n", encoding="utf-8")
-    (tmp_path / "log.md").write_text(LOG, encoding="utf-8")
-    wiki = tmp_path / "Wiki"
-    wiki.mkdir()
-    (wiki / "alpha-note.md").write_text(ALPHA, encoding="utf-8")
-    (wiki / "alpha-note-draft.md").write_text(DRAFT, encoding="utf-8")
-    (wiki / "gamma-orphan.md").write_text(GAMMA, encoding="utf-8")
-    return tmp_path
-
-
-def _run(vault: Path, *extra: str) -> dict:
-    proc = subprocess.run(
-        [sys.executable, str(SCRIPT), str(vault), "--scope", "all", "--json", *extra],
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-    assert proc.returncode == 0, proc.stderr
-    return json.loads(proc.stdout)
-
-
-def _pair(report: dict, a: str, b: str) -> dict:
-    return next(d for d in report["duplicate_titles"] if {d["a"], d["b"]} == {a, b})
 
 
 def test_scans_only_notes_and_reads_declared_types(vault):
@@ -182,118 +62,11 @@ def test_unindented_block_list_reads_as_filled(vault):
     assert tags == {"expected": 3, "present": 3, "filled": 3, "empty": 0, "empty_pct": 0.0, "exceeds": None}
 
 
-def test_raw_source_note_without_type_is_drift_checked_not_untyped(vault):
-    raw = vault / "Raw"
-    raw.mkdir()
-    (raw / "article.md").write_text(
-        "---\nsource_url: https://example.com/a\nsha256: " + "0" * 64 + "\n---\nraw text\n",
-        encoding="utf-8",
-    )
-    report = _run(vault)
-    assert report["untyped"] == []
-    by_path = {item["path"]: item for item in report["raw_sources"]}
-    assert by_path["Raw/article.md"]["status"] == "drift"
-
-
 def test_undeclared_fields_are_counted_against_schema(vault):
     entity = _run(vault)["types"]["entity"]
     assert entity["declared"] == ["aliases", "created", "tags", "type"]
     assert entity["variants"] == 1
     assert entity["undeclared"] == {"extra_field": {"present": 1, "present_pct": 100.0, "exceeds": None}}
-
-
-def test_orphan_graph(vault):
-    links = _run(vault)["links"]
-    assert links["orphans"] == ["Wiki/gamma-orphan.md"]
-    assert links["notes"]["Wiki/alpha-note.md"] == {"inbound": 1, "outbound": 1}
-    assert links["notes"]["Wiki/alpha-note-draft.md"] == {"inbound": 1, "outbound": 1}
-
-
-def test_typed_note_with_remote_source_and_no_raw_link_is_unresolved_not_drift(vault):
-    # A typed note's sha256 is the Raw Source's digest (B-4), never its own body's:
-    # hashing the synthesized body would report drift on every correct ingest.
-    by_path = {item["path"]: item for item in _run(vault)["raw_sources"]}
-    assert by_path["Wiki/alpha-note.md"] == {
-        "path": "Wiki/alpha-note.md",
-        "hashed_file": None,
-        "status": "unresolved",
-        "stored": "0" * 64,
-        "recomputed": None,
-    }
-    # An empty digest with nothing to backfill from is unresolved too, not unhashed.
-    assert by_path["Wiki/alpha-note-draft.md"]["status"] == "unresolved"
-
-
-RAW_BODY = "# Fed statement\n\nRates unchanged.\n"
-RAW_DIGEST = hashlib.sha256(RAW_BODY.encode("utf-8")).hexdigest()
-
-
-def _news_shaped_pair(vault: Path) -> Path:
-    raw_dir = vault / "Raw" / "2026-05-06"
-    raw_dir.mkdir(parents=True)
-    raw = raw_dir / "bloomberg-fed.md"
-    raw.write_text(
-        f"---\nsource_url: https://example.com/fed\nsha256: {RAW_DIGEST}\n---\n{RAW_BODY}",
-        encoding="utf-8",
-    )
-    (vault / "Wiki" / "fed-article.md").write_text(
-        "---\ntype: source\ncreated: 2026-05-06\n"
-        f"source_url: https://example.com/fed\nsha256: {RAW_DIGEST}\n---\n"
-        "# Fed\n\n## Lead\nRates held.\n\n## Source\n[[Raw/2026-05-06/bloomberg-fed]] (immutable original)\n",
-        encoding="utf-8",
-    )
-    return raw
-
-
-def test_typed_note_hashes_the_raw_source_its_body_links(vault):
-    raw = _news_shaped_pair(vault)
-    by_path = {item["path"]: item for item in _run(vault)["raw_sources"]}
-    assert by_path["Wiki/fed-article.md"] == {
-        "path": "Wiki/fed-article.md",
-        "hashed_file": "Raw/2026-05-06/bloomberg-fed.md",
-        "status": "fresh",
-        "stored": RAW_DIGEST,
-        "recomputed": RAW_DIGEST,
-    }
-    assert by_path["Raw/2026-05-06/bloomberg-fed.md"]["status"] == "fresh"
-    # A legacy empty digest on a note whose capture resolves is unhashed: backfill from it.
-    (vault / "Wiki" / "fed-article-legacy.md").write_text(
-        "---\ntype: source\nsource_url: https://example.com/fed\nsha256: \"\"\n---\n"
-        "## Source\n[[Raw/2026-05-06/bloomberg-fed]]\n",
-        encoding="utf-8",
-    )
-    legacy = {item["path"]: item for item in _run(vault)["raw_sources"]}["Wiki/fed-article-legacy.md"]
-    assert legacy["status"] == "unhashed" and legacy["recomputed"] == RAW_DIGEST
-    # --skip Raw takes the capture out of the note set, not out of the drift
-    # check, and its link is still provenance rather than a wiki edge.
-    skipped = _run(vault, "--skip", "Raw")
-    by_path = {item["path"]: item for item in skipped["raw_sources"]}
-    assert by_path["Wiki/fed-article.md"]["status"] == "fresh"
-    assert by_path["Raw/2026-05-06/bloomberg-fed.md"]["status"] == "fresh"
-    assert "Wiki/fed-article.md" in skipped["links"]["orphans"]
-    # The publisher edits the article: the re-captured body drifts from the pair.
-    raw.write_text(raw.read_text(encoding="utf-8") + "\nCorrection appended.\n", encoding="utf-8")
-    by_path = {item["path"]: item for item in _run(vault)["raw_sources"]}
-    assert by_path["Wiki/fed-article.md"]["status"] == "drift"
-    assert by_path["Wiki/fed-article.md"]["hashed_file"] == "Raw/2026-05-06/bloomberg-fed.md"
-
-
-def test_raw_sources_are_drift_checked_but_never_scope_subjects(vault):
-    raw = vault / "Raw"
-    raw.mkdir()
-    (raw / "alpha-capture.md").write_text(
-        "---\ntitle: Alpha\nsource_url: https://example.com/alpha\nsha256: " + "0" * 64 + "\n---\nraw text\n",
-        encoding="utf-8",
-    )
-    newest = max(p.stat().st_mtime for p in (vault / "Wiki").iterdir()) + 10
-    os.utime(raw / "alpha-capture.md", (newest, newest))
-    report = _run(vault, "--scope", "recent:1")
-    # The newest file is the capture, yet the one subject is a wiki note.
-    assert report["notes"] == {"total": 3, "raw_sources": 1, "in_scope": 1}
-    assert list(report["links"]["notes"]) != ["Raw/alpha-capture.md"]
-    assert {item["path"] for item in report["raw_sources"]} >= {"Raw/alpha-capture.md"}
-    # A capture titled like its note is not a duplicate-title candidate.
-    assert not any("Raw/" in d["a"] or "Raw/" in d["b"] for d in _run(vault)["duplicate_titles"])
 
 
 def test_audit_reports_are_never_scope_subjects(vault):
@@ -308,70 +81,10 @@ def test_audit_reports_are_never_scope_subjects(vault):
     assert "audit-report" not in result["types"]
 
 
-def test_wikilinks_inside_code_are_examples_not_links(vault):
-    (vault / "Wiki" / "gamma-orphan.md").write_text(
-        # A double-backtick span, and a fence that opens with ``` and closes with
-        # ```` — a longer delimiter CommonMark accepts.
-        GAMMA + "\nExample: `[[alpha-note]]`, ``[[alpha-note]]`` and\n\n```\n[[alpha-note-draft]]\n````\n",
-        encoding="utf-8",
-    )
-    links = _run(vault)["links"]
-    assert "Wiki/gamma-orphan.md" in links["orphans"]
-    assert links["notes"]["Wiki/alpha-note.md"]["inbound"] == 1
-
-
 def test_log_entries_naming_missing_files(vault):
     log = _run(vault)["log"]
     assert log["entries"] == 3
     assert log["missing_files"] == [{"line": 4, "file": "Wiki/deleted-note.md"}]
-
-
-def test_raw_captures_and_attachment_embeds_are_not_edges(vault):
-    raw = vault / "Raw"
-    raw.mkdir()
-    # Captured text mentioning the orphan is the source's link, not the vault's.
-    (raw / "capture.md").write_text(
-        "---\nsource_url: https://example.com/c\nsha256: " + "0" * 64 + "\n---\nSee [[Wiki/gamma-orphan]] and [[alpha-note]].\n",
-        encoding="utf-8",
-    )
-    # Attachments of any extension are embeds; a dotted note name still resolves.
-    (vault / "Wiki" / "Node.js.md").write_text("---\ntype: entity\n---\nabout node\n", encoding="utf-8")
-    (vault / "Wiki" / "gamma-orphan.md").write_text(
-        GAMMA + "\n![[recording.mp3]] ![[clip.webp]] ![[deck.pdf]]\n", encoding="utf-8"
-    )
-    (vault / "Wiki" / "alpha-note-draft.md").write_text(DRAFT + "\nRuntime: [[Node.js]].\n", encoding="utf-8")
-    links = _run(vault)["links"]
-    assert "Wiki/gamma-orphan.md" in links["orphans"]
-    assert links["notes"]["Wiki/alpha-note.md"]["inbound"] == 1
-    assert links["notes"]["Wiki/Node.js.md"]["inbound"] == 1
-    assert links["notes"]["Wiki/alpha-note-draft.md"]["outbound"] == 2
-
-
-def test_links_in_yaml_comments_are_not_edges_but_links_in_values_are(vault):
-    (vault / "Wiki" / "gamma-orphan.md").write_text(
-        GAMMA.replace("extra_field: yes\n", "extra_field: yes\nrelated: []   # e.g. [[alpha-note]]\n"),
-        encoding="utf-8",
-    )
-    links = _run(vault)["links"]
-    assert "Wiki/gamma-orphan.md" in links["orphans"]
-    assert links["notes"]["Wiki/alpha-note.md"]["inbound"] == 1
-    (vault / "Wiki" / "gamma-orphan.md").write_text(
-        GAMMA.replace("extra_field: yes\n", "extra_field: yes\nrelated: \"[[alpha-note]]\"\n"),
-        encoding="utf-8",
-    )
-    links = _run(vault)["links"]
-    assert links["notes"]["Wiki/gamma-orphan.md"]["outbound"] == 1
-    assert links["notes"]["Wiki/alpha-note.md"]["inbound"] == 2
-
-
-def test_self_links_under_any_spelling_are_not_edges(vault):
-    (vault / "Wiki" / "gamma-orphan.md").write_text(
-        GAMMA + "\nSelf: [[gamma-orphan]] and [[Wiki/gamma-orphan]] and [[gamma-orphan#Heading|me]].\n",
-        encoding="utf-8",
-    )
-    links = _run(vault)["links"]
-    assert links["notes"]["Wiki/gamma-orphan.md"] == {"inbound": 0, "outbound": 0}
-    assert "Wiki/gamma-orphan.md" in links["orphans"]
 
 
 def test_fenced_examples_under_the_taxonomy_declare_nothing(vault):
@@ -388,56 +101,6 @@ def test_fenced_examples_under_the_taxonomy_declare_nothing(vault):
     assert report["tags"]["ghost"] == {"count": 1, "declared": False, "exceeds": True}
 
 
-def test_ambiguous_bare_links_resolve_by_folder_or_not_at_all(vault):
-    for folder in ("A", "B"):
-        (vault / folder).mkdir()
-        (vault / folder / "foo.md").write_text(f"---\ntype: entity\n---\n{folder} foo\n", encoding="utf-8")
-    (vault / "B" / "source.md").write_text("---\ntype: entity\n---\nSee [[foo]].\n", encoding="utf-8")
-    (vault / "Wiki" / "gamma-orphan.md").write_text(GAMMA + "\nAlso [[foo]].\n", encoding="utf-8")
-    links = _run(vault)["links"]
-    assert links["notes"]["B/foo.md"]["inbound"] == 1      # same folder as B/source.md
-    assert links["notes"]["A/foo.md"]["inbound"] == 0      # not handed the link by sort order
-    assert "A/foo.md" in links["orphans"]
-    assert links["notes"]["Wiki/gamma-orphan.md"]["outbound"] == 1   # ambiguous from Wiki/: unresolved, still a link
-    assert "Wiki/gamma-orphan.md" not in links["orphans"]
-    # The same-folder rule also decides the attachment and provenance prefilters:
-    # a dotted name shared by two folders, and a bare name a capture also has.
-    for folder in ("A", "B"):
-        (vault / folder / "Node.js.md").write_text("---\ntype: entity\n---\nnode\n", encoding="utf-8")
-    (vault / "Raw").mkdir()
-    (vault / "Raw" / "foo.md").write_text("---\nsource_url: https://x/foo\nsha256: " + "0" * 64 + "\n---\ncapture\n", encoding="utf-8")
-    (vault / "B" / "source.md").write_text("---\ntype: entity\n---\nSee [[foo]] and [[Node.js]].\n", encoding="utf-8")
-    links = _run(vault)["links"]
-    assert links["notes"]["B/Node.js.md"]["inbound"] == 1
-    assert links["notes"]["B/foo.md"]["inbound"] == 1
-    assert links["notes"]["B/source.md"]["outbound"] == 2
-    # With a wiki `foo` in two folders and a capture `foo`, the ambiguous link
-    # from Wiki/ is still a wiki link (unresolved), not provenance.
-    assert links["notes"]["Wiki/gamma-orphan.md"]["outbound"] == 1
-
-
-def test_provenance_link_to_a_capture_is_not_a_wiki_relationship(vault):
-    _news_shaped_pair(vault)
-    # The Article's only link is its `## Source` line: no wiki relationship, so
-    # it is an orphan, while its provenance still resolves for the drift check.
-    report = _run(vault)
-    assert "Wiki/fed-article.md" in report["links"]["orphans"]
-    assert {i["path"]: i["status"] for i in report["raw_sources"]}["Wiki/fed-article.md"] == "fresh"
-    # A long attachment extension and a code span crossing a line break are
-    # embeds and examples too.
-    (vault / "Wiki" / "gamma-orphan.md").write_text(
-        GAMMA + "\n![[diagram.excalidraw]] and ``literal\n[[alpha-note]]`` here.\n"
-        "\n> [!note]\n> ```md\n> [[alpha-note]] inside a callout fence\n>```\n\nAfter the callout: [[Node.js]].\n",
-        encoding="utf-8",
-    )
-    # The callout fence closes on `>``` ` (no space), and text after it is text again.
-    (vault / "Wiki" / "Node.js.md").write_text("---\ntype: entity\n---\nnode\n", encoding="utf-8")
-    links = _run(vault)["links"]
-    assert "Wiki/gamma-orphan.md" not in links["orphans"]
-    assert links["notes"]["Wiki/gamma-orphan.md"]["outbound"] == 1
-    assert links["notes"]["Wiki/alpha-note.md"]["inbound"] == 1
-
-
 def test_log_path_tokens_are_not_satisfied_by_a_basename_elsewhere(vault):
     # `Wiki/deleted-note.md` stays missing when only `Archive/deleted-note.md`
     # exists; a bare basename in the log still matches any file of that name.
@@ -449,12 +112,15 @@ def test_log_path_tokens_are_not_satisfied_by_a_basename_elsewhere(vault):
         log.write("## [2026-05-11] schema | bump to v1.2\n")
         log.write("## [2026-05-12] create | entity | Wiki/My Note.md\n")
         log.write("## [2026-05-13] query | summarize Wiki/alpha-note.md and Wiki/gone.md\n")
+        log.write("## [2026-05-14] schema | updated SCHEMA.md\n")
+        log.write("## [2026-05-15] create | entity | ./My Root Note.md\n")
     report = _run(vault)["log"]
-    assert report["entries"] == 8
+    assert report["entries"] == 10
     assert report["missing_files"] == [
         {"line": 4, "file": "Wiki/deleted-note.md"},
         {"line": 7, "file": "Raw/recording.mp3"},
         {"line": 9, "file": "Wiki/My Note.md"},
+        {"line": 12, "file": "./My Root Note.md"},
     ]
 
 
@@ -530,9 +196,6 @@ def test_skip_drops_a_folder_and_bad_scope_is_rejected(vault):
     assert "scope must be" in proc.stderr
 
 
-PRESETS = Path(__file__).resolve().parents[2] / "skills" / "core" / "maintaining-obsidian" / "presets"
-
-
 @pytest.mark.parametrize("preset,type_name", [("llm-wiki", "source"), ("news", "article")])
 def test_preset_source_types_declare_the_provenance_pair(tmp_path, preset, type_name):
     # obsidian.md B-4: the typed note carries source_url and sha256; a preset whose
@@ -561,19 +224,6 @@ def test_preset_taxonomies_declare_their_top_level_tags(tmp_path, preset, type_n
     assert tag.split("/")[0] in report["schema"]["declared_tags"]
     assert report["tags"][tag] == {"count": 1, "declared": True, "exceeds": False}
     assert report["tags"]["made-up"] == {"count": 1, "declared": False, "exceeds": True}
-
-
-GENERIC_SOURCE = (
-    "---\ntype: source\ncreated: 2026-05-06\nlangs: [en, zh]\n"
-    "source_url: https://example.com/{n}\nsource_author: Someone\nsha256: {digest}\n"
-    "tags: [source]\naliases: []\n---\nbody\n"
-)
-PAPER_SOURCE = (
-    "---\ntype: source\ncreated: 2026-05-06\nlangs: [en, zh]\n"
-    "source_url: https://arxiv.org/abs/1\nsource_author: [A, B]\nsha256: {digest}\n"
-    "venue: NeurIPS\nyear: 2025\nmethodology: ''\nreading_status: queued\n"
-    "cites: []\ncited_by: []\ntags: [paper]\naliases: []\n---\nbody\n"
-)
 
 
 def test_variant_only_fields_are_expected_of_the_notes_fitting_that_variant(tmp_path):
@@ -651,19 +301,6 @@ def test_variant_fences_under_subheadings_keep_their_type_section(tmp_path):
     assert "venue" not in source["fields"]
 
 
-def test_bare_provenance_link_colliding_with_a_wiki_name_is_unresolved(vault):
-    raw = vault / "Raw"
-    raw.mkdir()
-    (raw / "foo.md").write_text("---\nsource_url: https://x/foo\nsha256: " + "0" * 64 + "\n---\ncapture\n", encoding="utf-8")
-    (vault / "Wiki" / "foo.md").write_text("---\ntype: entity\n---\nwiki foo\n", encoding="utf-8")
-    (vault / "Wiki" / "typed.md").write_text(
-        "---\ntype: source\nsource_url: https://x/t\nsha256: " + "0" * 64 + "\n---\nSee [[foo]].\n", encoding="utf-8"
-    )
-    by_path = {i["path"]: i for i in _run(vault)["raw_sources"]}
-    assert by_path["Wiki/typed.md"]["status"] == "unresolved"
-    assert by_path["Wiki/typed.md"]["hashed_file"] is None
-
-
 def test_a_heading_naming_the_type_as_a_substring_is_still_a_base(tmp_path):
     # `## Catalog defaults` contains "log" but does not name the type `log`.
     (tmp_path / "SCHEMA.md").write_text(
@@ -678,26 +315,6 @@ def test_a_heading_naming_the_type_as_a_substring_is_still_a_base(tmp_path):
     types = _run(tmp_path)["types"]
     assert types["log"]["variants"] == 1 and types["log"]["fields"]["created"]["empty"] == 1
     assert types["daily-aggregate"]["variants"] == 1
-
-
-def test_links_with_a_path_resolve_by_that_path_only(vault):
-    raw = _news_shaped_pair(vault)
-    # The Article's link names the wrong day: the only bloomberg-fed.md lives
-    # under 2026-05-06, and it must not stand in for the missing capture.
-    article = vault / "Wiki" / "fed-article.md"
-    article.write_text(
-        article.read_text(encoding="utf-8").replace("[[Raw/2026-05-06/bloomberg-fed]]", "[[Raw/2026-05-07/bloomberg-fed]]"),
-        encoding="utf-8",
-    )
-    report = _run(vault)
-    assert {i["path"]: i["status"] for i in report["raw_sources"]}["Wiki/fed-article.md"] == "unresolved"
-    assert raw.exists()
-    # The same rule in the link graph: a path link to a note that is not there
-    # gives the same-named note elsewhere no backlink.
-    (vault / "Wiki" / "gamma-orphan.md").write_text(GAMMA + "\nSee [[Archive/alpha-note]].\n", encoding="utf-8")
-    links = _run(vault)["links"]["notes"]
-    assert links["Wiki/alpha-note.md"]["inbound"] == 1
-    assert links["Wiki/gamma-orphan.md"]["outbound"] == 1
 
 
 def test_minimal_preset_placeholders_declare_no_types_or_tags(tmp_path):
@@ -723,3 +340,4 @@ def test_minimal_preset_placeholders_declare_no_types_or_tags(tmp_path):
         schema.write("\nIndented example:\n\n    ```yaml\n    type: ghost\n    ```\n")
         schema.write("\n> [!example]\n> ```yaml\n> type: ghost\n> ```\n")
     assert _run(tmp_path)["schema"]["declared_types"] == ["book"]
+
